MB-44253: Remove try/catch from ThreadPoolExecutor::runTask
This try/catch block prevents unhandled exceptions in GlobalTasks from
triggering std::terminate, and in turn generating a Breakpad minidump
with the state of the process when the exception was thrown (as
opposed to where it was caught).

Started a discussion with Folly about a more generic way to address
this (i.e. making the try/catch somehow optional) - see
https://github.com/facebook/folly/issues/1525 - but in the short-term
simply removing the code from our branch.